wip on restructuring and renaming files in the project
