+++ scene/src/components/arrow-toast/index.ts
@@ -1,0 +1,1 @@
+export { default as ArrowToast } from './ArrowToast';
