+++ scene/src/components/friend-item/friendItem.tsx
