+++ scene/src/components/slider/slider.tsx
