+++ scene/src/utils/ui-utils.ts
@@ -88,6 +88,7 @@ export function isValidURL(url: string): boolean {
   const urlRegex = /^(https?:\/\/)?([\da-z.-]+)\.([a-z.]{2,6})([/\w .-]*)*\/?$/
   return urlRegex.test(url)
 }
+
 export function isValidDate(date: string): boolean {
   const dateRegex = /^(0[1-9]|[12][0-9]|3[01])\/(0[1-9]|1[0-2])\/\d{4}$/
   return dateRegex.test(date)
@@ -141,5 +142,34 @@ export function sliderPercentageToValue(
     return Math.round(value)
   } else {
     return value
+  }
+}
+
+export function formatEventTime(timestampInSecs: number): string {
+  const now = new Date();
+  const eventDate = new Date(timestampInSecs * 1000); // convert to miliseconds
+
+  if (eventDate > now) {
+      // Format: MON, 24 JUL AT 02:00PM
+      const days = ["SUN", "MON", "TUE", "WED", "THU", "FRI", "SAT"];
+      const months = ["JAN", "FEB", "MAR", "APR", "MAY", "JUN", "JUL", "AUG", "SEP", "OCT", "NOV", "DEC"];
+      const day = days[eventDate.getUTCDay()];
+      const date = eventDate.getUTCDate();
+      const month = months[eventDate.getUTCMonth()];
+      const hours = eventDate.getUTCHours();
+      const minutes = eventDate.getUTCMinutes();
+      const formattedTime = `${isNaN(hours % 12) ? 12 : hours % 12}:${minutes.toString().padStart(2, "0")}${hours >= 12 ? "PM" : "AM"}`;
+      return `${day}, ${date} ${month} AT ${formattedTime}`;
+  } else {
+      // Calcular diferencia
+      const diff = Math.abs(now.getTime() - eventDate.getTime()); // Diferencia en milisegundos
+      const minutesDiff = Math.floor(diff / (1000 * 60));
+      const hoursDiff = Math.floor(minutesDiff / 60);
+
+      if (hoursDiff > 0) {
+          return `Event started ${hoursDiff} hours ago.`;
+      } else {
+          return `Event started ${minutesDiff} minutes ago.`;
+      }
   }
 }
